backend/requirements.txt
# ==================== SPECTER Legal Assistant Dependencies ====================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
motor==3.5.1
pymongo==4.6.3

# Authentication & Security
python-jose[cryptography]==3.5.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6

# Document Processing
python-docx==1.1.0
pypdf<5
fonttools>=4.33.3
Pillow>=9.2.0

# Vector Search & AI
sentence-transformers>=2.7.0
chromadb>=0.5.0
numpy<2.0

# HTTP Requests
requests==2.31.0
httpx>=0.24.1,<0.26.0

# Payment Processing
razorpay==1.4.1

# Email Support
secure-smtplib==0.1.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Production Server
gunicorn==21.2.0

# Monitoring & Logging
structlog==23.2.0
